requirements.txt
click==6.2
dnspython==1.12.0
lxml==3.5.0
pytz==2015.7
PyYAML==3.11
requests==0.14.1
route53==1.0
wheel==0.24.0
